refactor: split gateway router/server.go by responsibility cluster (M3-9, ARCH-4)
Pure file-move refactor. No behavior change. No renames.

Files created:
- oauth.go: GitHub OAuth flow, demo-session, API-key rotate/revoke
- disclosure.go: tier1/tier2 disclosure structs + synthesis funcs + routing metadata cache
- chat_proxy.go: handleChatCompletions, SSE re-emit, quota-settle path
- admin.go: /admin/kill-switch, /admin/capacity-*, /admin/feedback-summary + helpers
- auth_helpers.go: operator bearer auth, requireBearer, clientIP, trusted-proxy IP logic

server.go: 2519 -> 937 lines (struct, route table, status/poolz cluster, shared helpers)

go test ./internal/router/... -count=2: green
go vet ./...: clean
